✅  Update: Refactor user hooks to use constants for query keys and improve error handling in API calls
