+++ src/hooks/index.ts
@@ -1,4 +1,3 @@
-//hooks user
 export * from './user/useCreateUserHook';
 export * from './user/useDeleteUserHook';
 export * from './user/useUpdateUserHook';
