+++ src/hooks/user/useDeleteUserHook.ts
@@ -1,22 +1,17 @@
 import { useMutation, useQueryClient } from '@tanstack/react-query';
 import { userApi } from '../../lib/api';
+import { QUERY_KEYS } from '../../lib/api/constants';
 
 const deleteUser = async (id: string): Promise<void> => {
-  try {
-    await userApi.delete(`/user/${id}`);
-  } catch (error) {
-    console.error(`Error eliminando usuario con ID ${id}:`, error);
-    throw error;
-  }
+  await userApi.delete(`/user/${id}`);
 };
 
 export function useDeleteUser() {
   const queryClient = useQueryClient();
-
   return useMutation<void, Error, string>({
     mutationFn: deleteUser,
     onSuccess: () => {
-      queryClient.invalidateQueries({ queryKey: ['users'] });
+      queryClient.invalidateQueries({ queryKey: [QUERY_KEYS.USERS] });
     },
   });
 }
