+++ src/hooks/user/useGetAllUsersHook.ts
@@ -1,6 +1,7 @@
 import { useQuery } from '@tanstack/react-query';
 import { userApi } from '../../lib/api';
-import { User } from '../../types';  
+import { User } from '../../types/userType';
+import { QUERY_KEYS } from '../../lib/api/constants';
 
 const fetchUsers = async (): Promise<User[]> => {
   const { data } = await userApi.get<User[]>('/user');
@@ -9,7 +10,7 @@ const fetchUsers = async (): Promise<User[]> => {
 
 export function useGetAllUsers() {
   return useQuery({
-    queryKey: ['users'],
+    queryKey: [QUERY_KEYS.USERS],
     queryFn: fetchUsers,
   });
 }
