+++ src/hooks/user/useGetUserByIdHook.ts
@@ -1,6 +1,6 @@
 import { useQuery } from '@tanstack/react-query';
 import { userApi } from '../../lib/api';
-import { User } from '../../types';  
+import { User } from '../../types/userType';
 
 const fetchUserById = async (id: string): Promise<User> => {
   const { data } = await userApi.get<User>(`/user/${id}`);
@@ -8,11 +8,9 @@ const fetchUserById = async (id: string): Promise<User> => {
 };
 
 export function useGetUserById(id: string) {
-  const queryKey = ['user', id];
   return useQuery({
-    queryKey,
+    queryKey: ['user', id],
     queryFn: () => fetchUserById(id),
-    enabled: !!id, 
+    enabled: !!id,
   });
 }
-
