+++ src/lib/api/constants.ts
@@ -1,0 +1,3 @@
+export const QUERY_KEYS = {
+    USERS: 'users',
+};
